requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
crashtest==0.4.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.7
dulwich==0.22.8
easyocr==1.7.1
ecdsa==0.19.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.12
fastjsonschema==2.19.1
filelock==3.13.1
findpython==0.6.3
fqdn==1.5.1
fsspec==2024.6.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
imageio==2.34.2
inflection==0.5.1
installer==0.7.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keyring==25.6.0
lazy_loader==0.4
Mako==1.3.9
Markdown==3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.10.2.4
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pbs-installer==2025.2.12
pillow==11.1.0
pipx==1.7.1
pkginfo==1.12.1.2
platformdirs==3.11.0
poetry==2.1.1
poetry-core==2.1.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.4.8
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.2.0
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
RapidFuzz==3.12.2
referencing==0.35.1
requests==2.32.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9
scikit-image==0.24.0
scipy==1.14.0
Send2Trash==1.8.3
setuptools==69.0.1
shapely==2.0.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.2
terminado==0.18.1
tifffile==2024.8.10
tinycss2==1.3.0
tomlkit==0.13.2
torch==2.4.0
torchvision==0.19.0
tornado==6.4
traitlets==5.14.3
trove-classifiers==2025.3.3.18
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
userpath==1.9.2
uvicorn==0.34.0
validate-docbr==1.10.0
virtualenv==20.29.3
voila==0.5.7
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.41.3
zstandard==0.23.0
